% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barley_data.R
\docType{data}
\name{barley_map}
\alias{barley_map}
\title{Genetic Map of Barley SNP Markers}
\format{A list with 7 elements:
\describe{
  \item{Chromosome 1}{198 markers}
  \item{Chromosome 2}{259 markers}
  \item{Chromosome 3}{212 markers}
  \item{Chromosome 4}{219 markers}
  \item{Chromosome 5}{262 markers}
  \item{Chromosome 6}{216 markers}
  \item{Chromosome 7}{224 markers}
}}
\source{
\url{https://triticeaetoolbox.org/barley/}
}
\usage{
barley_map
}
\description{
A \code{list} of genetic map positions of 1590 bi-allelic SNP markers on
the 7 barley chromosomes. Can be used as a genetic map in the package
\code{qtl}.
}
\keyword{datasets}

