% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barley_data.R
\docType{data}
\name{barley_SNPs}
\alias{barley_SNPs}
\title{Barley Genotype Data}
\format{A matrix with 764 rows and 1590 columns, where each row is a breeding
line and each column is a bi-allelic SNP marker. Markers are encoded as
z {0, 1, 2} where z is the number of "A" alleles.}
\source{
\url{https://triticeaetoolbox.org/barley/}
}
\usage{
barley_SNPs
}
\description{
A \code{matrix} of genotype data for barley breeding lines from
the University of Minnesota and North Dakota State University.
}
\keyword{datasets}

