% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsimpute.R
\name{fsimpute}
\alias{fsimpute}
\title{Impute genotypes in bi-parental populations of finite selfing}
\usage{
fsimpute(cross, founders, finals, prob.threshold = 0.7, step = 0,
  off.end = 0, error.prob = 1e-04, map.function = c("haldane", "kosambi",
  "c-f", "morgan"), stepwidth = c("fixed", "variable", "max"))
}
\arguments{
\item{cross}{An object of class \code{cross} from the package \code{qtl}. Genotype
data should only include those of the progeny in the family.}

\item{founders}{A list of founder genotypes, where each element in the list
is a \code{matrix} of genotypes on a single chromosome. Genotypes should be
coded as \code{z {0, 1, 2, NA}} where \code{z} is the number of reference alleles.
Column names should be marker names. This is the observed genotype data.}

\item{finals}{A list of progeny genotypes. The encoding and formatting should
be identical to the argument \code{founders}.}

\item{prob.threshold}{The minimum conditional genotype probability to declare
a site as having been inherited from a certain parent.}

\item{step}{See \code{\link[qtl]{calc.genoprob}}.}

\item{off.end}{See \code{\link[qtl]{calc.genoprob}}.}

\item{error.prob}{See \code{\link[qtl]{calc.genoprob}}.}

\item{stepwidth}{See \code{\link[qtl]{calc.genoprob}}.}

\item{map.functions}{See \code{\link[qtl]{calc.genoprob}}.}
}
\value{
The \code{cross} object is returned with the additional item \code{imputed},
which contains the following information:
\describe{
  \item{$stats}{Some measures on the imputed genotypes, including the proportion of missing data}
  \item{$founders}{The imputed founder genotypes}
  \item{$finals}{The imputed progeny genotypes}
  }
}
\description{
Imputes missing genotype data using conditional genotype probabilities
derived from a Hidden Markov Model. Founder genotypes are then imputed
from the conditional genotype probabilities in the progeny. Then progeny
genotypes are imputed using the imputed founder genotypes.
}

