% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsimpute.R
\name{fsimpute}
\alias{fsimpute}
\title{Impute genotypes in bi-parental populations of finite selfing}
\usage{
fsimpute(cross, prob.threshold = 0.7, type = "discrete", step = 0,
  off.end = 0, error.prob = 1e-04, map.function = c("haldane", "kosambi",
  "c-f", "morgan"), stepwidth = c("fixed", "variable", "max"))
}
\arguments{
\item{cross}{An object of class \code{cross} from the package
\code{\link[qtl]{read.cross}}. This object should be generated using the
function \code{\link[fsimpute]{prep_cross}}.}

\item{prob.threshold}{The minimum conditional genotype probability to declare
a site as having been inherited from a certain parent.}

\item{type}{The genotype output type. Can be \code{discrete} to code each
imputed genotype as the most likely genotype, or \code{continuous} to code
each imputed genotype as a weighted average based on the genotype probabilities.}

\item{step}{See \code{\link[qtl]{calc.genoprob}}.}

\item{off.end}{See \code{\link[qtl]{calc.genoprob}}.}

\item{error.prob}{See \code{\link[qtl]{calc.genoprob}}.}

\item{stepwidth}{See \code{\link[qtl]{calc.genoprob}}.}

\item{map.functions}{See \code{\link[qtl]{calc.genoprob}}.}
}
\value{
The \code{cross} object is returned with the additional item \code{imputed},
which contains the following information:
\describe{
  \item{$stats}{Some measures on the imputed genotypes, including the proportion of missing data}
  \item{$founders}{The imputed founder genotypes}
  \item{$finals}{The imputed progeny genotypes}
  }
}
\description{
Imputes missing genotype data using conditional genotype probabilities
derived from a Hidden Markov Model. Founder genotypes are then imputed
from the conditional genotype probabilities in the progeny. Then progeny
genotypes are imputed using the imputed founder genotypes.
}

